collapse card encodings to a single location encoding
Drop the index encoding and the GLITCH suit; every (entry, suit) pair is now a
card at location = entry + 14*suit (0..55, 56 cards). Jokers become distinct
suited cards (HEARTS=28, SPADES=42); yield=(JOKER,CLUBS)=0 and
resign=(JOKER,DIAMONDS)=14 are sentinels that are never dealt (added
Card::isYield/isResign). getPower now returns JOKER_NERF for any joker, never a
suit power. The phaseinfo string and pickle serialize locations.

Combo bitwise is now a u64 bitmask over card locations (canonical combo
identity), computed in loadDetails and fillComboEntry.

Regenerate the get_expansion_at golden snapshots and update the encoding
characterization tests. Migrate the suitpref/bin2df bridges from Suit.GLITCH to
Entry.JOKER.

+++ src/regi_py/core/combo.h
@@ -11,7 +11,7 @@ namespace regi
        private:
         i32 baseDmg;
         u32 powers;
-        u32 bitrep;
+        u64 bitrep;
 
        public:
         std::vector<Card> parts;
@@ -23,8 +23,8 @@ namespace regi
         i32 getBaseDefense() const;
         friend std::ostream &operator<<(std::ostream &, const Combo &);
         //
-        void setBitrep(u32);
-        u32 getBitrep() const;
+        void setBitrep(u64);
+        u64 getBitrep() const;
     };
 
 } /* namespace regi */
